Co-locate all relevant code
