+++ apps/docs/app/entry.server.tsx
@@ -1,10 +1,10 @@
 import { CacheProvider } from "@emotion/react";
 import createEmotionServer from "@emotion/server/create-instance";
-import { renderToString } from "react-dom/server";
 import type { EntryContext } from "@remix-run/node";
 import { RemixServer } from "@remix-run/react";
-import { createEmotionCache } from "./features/chakra-setup/createEmotionCache";
-import { ServerStyleContext } from "./features/chakra-setup/styleContext";
+import { renderToString } from "react-dom/server";
+import { createEmotionCache } from "./features/setup/chakra-setup/createEmotionCache";
+import { ServerStyleContext } from "./features/setup/chakra-setup/styleContext";
 
 export default function handleRequest(
   request: Request,
