+++ apps/docs/app/features/illustrations/NotFoundIllustration.tsx
