+++ apps/docs/app/features/portable-text/ImageWithCaption.tsx
