+++ apps/docs/app/features/setup/error-boundary/RootErrorBoundary.tsx
