+++ apps/docs/app/routes/_base.ressurser.design-tokens/AnimationTokens.tsx
@@ -12,7 +12,7 @@ import {
   Thead,
   Tr,
 } from "@vygruppen/spor-react";
-import { useTokenFormatter } from "~/features/tokens/useTokenFormatter";
+import { useTokenFormatter } from "~/routes/_base.ressurser.design-tokens/useTokenFormatter";
 import { SharedTokenLayout } from "./SharedTokenLayout";
 
 export function AnimationTokens(props: BoxProps) {
