+++ apps/docs/app/routes/_base.ressurser.design-tokens/TypographyTokens.tsx
@@ -13,8 +13,8 @@ import {
   Tr,
 } from "@vygruppen/spor-react";
 import { Fragment } from "react";
-import { LinkableHeading } from "~/features/linkable-heading/LinkableHeading";
-import { useTokenFormatter } from "~/features/tokens/useTokenFormatter";
+import { LinkableHeading } from "~/features/portable-text/LinkableHeading";
+import { useTokenFormatter } from "~/routes/_base.ressurser.design-tokens/useTokenFormatter";
 import { SharedTokenLayout } from "./SharedTokenLayout";
 
 type TypographyToken = {
