+++ apps/docs/app/routes/_base.ressurser.design-tokens/ZIndexTokens.tsx
@@ -11,7 +11,7 @@ import {
   Thead,
   Tr,
 } from "@vygruppen/spor-react";
-import { useTokenFormatter } from "~/features/tokens/useTokenFormatter";
+import { useTokenFormatter } from "~/routes/_base.ressurser.design-tokens/useTokenFormatter";
 import { SharedTokenLayout } from "./SharedTokenLayout";
 
 export function ZIndexTokens(props: BoxProps) {
