+++ apps/docs/app/routes/_base/content-menu/GlobalSearchInput.tsx
