+++ apps/docs/app/routes/_base/content-menu/SearchableContentMenu.tsx
