+++ apps/docs/app/routes/_base/table-of-contents/useScrollSpy.ts
