+++ apps/docs/app/routes/_index/HeroSection.tsx
